Version 2.0.0

+++ NSColorExtended.h
@@ -19,7 +19,7 @@
 +(NSColor*) colorWithRgbaString:(NSString*)string;
 -(NSString*) rgbaString;
 
--(float) grayLevel;
+-(CGFloat) grayLevel;
 -(BOOL)  isRGBEqualTo:(NSColor*)other;
 
 @end
